A little more clean up
