+++ fabfile/app.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
 from configure import ConfigTask
 from fabric.api import sudo, env, cd, task
