+++ fabfile/dev.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from configure import ConfigTask
 from fabric.api import local, task, sudo, cd, env
 
 
@@ -10,7 +13,7 @@ def rs(port=8000):
 
 
 @task
-def git_pull():
+def pull():
     """
     Pull the lastest changes from the GitHub repo
     """
@@ -19,3 +22,16 @@ def git_pull():
             'git pull origin master',
             user=env.app_user
         )
+
+
+@task(task_class=ConfigTask)
+def ssh(ec2_instance=''):
+    """
+    Log into the EC2 instance using SSH.
+    """
+    if not ec2_instance:
+        try:
+            ec2_instance = env.hosts[0]
+        except IndexError:
+            ec2_instance = require_input('EC2 Host [Required]:')
+    local("ssh %s@%s -i %s" % (env.user, ec2_instance, env.key_filename[0]))
